pyproject.toml
[project]
name = "ai-birthday-calendar"
version = "1.0.0"
description = "Self-hosted birthday tracker with AI-powered gift suggestions and email reminders"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "apscheduler>=3.10.4",
    "openai>=1.10.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.25.0",
    "ruff>=0.4.0",
]
